+++ src/index.ts
@@ -1,7 +1,5 @@
 export * from './common';
 export * from './create-async';
 export * from './use-async';
-export * from './asyncController';
-export * from './useAsyncFunction';
 export * from './share-loading';
 export * from './take-latest-promise';